updated the return json, now nulls are not represented in json format. instead of returning subscribeconfirmations now I deal with subscription instead.

+++ Backend/Project/DTOs/SubscriptionResponseDTO.cs
@@ -4,6 +4,6 @@ public class SubscriptionResponseDTO
 {
 
     public double Price { get; set; }
-    public double DaysLeft { get; set; }
+    public int DaysLeft { get; set; }
     public string StreamingServiceName { get; set; }
 }
